move message related code to message module, note type to note module
